remove app.js #2691
