+++ assets/services/search_filter_form.js
@@ -1,0 +1,15 @@
+document?.querySelectorAll('.select-search-filter-form')?.forEach(select => {
+    select.addEventListener(
+        "change",
+        () => {
+            setBadge(select);
+        },
+        false,
+      );
+})
+
+document?.querySelectorAll('[data-removable="true"]')?.forEach(removale => {
+    removale.addEventListener('click', () => {
+        removeBadge(removale);
+    })
+})
