+++ templates/base.html.twig
@@ -130,7 +130,6 @@
 
 <script type="module" src="{{ asset('build/dsfr/dsfr.module.min.js') }}"></script>
 <script nomodule src="{{ asset('build/dsfr/dsfr.nomodule.min.js') }}"></script>
-<script src="{{ asset('js/app.js?v='~''|date('YmdH')) }}"></script>
 
 {% if 'back_cartographie' in app.request.get('_route') %}
     <script src="https://cdn.jsdelivr.net/npm/leaflet@1.7.1/dist/leaflet.js"></script>
